requirements.txt
# Entorno del proyecto (notebook + app Streamlit)
numpy>=1.24
matplotlib>=3.7
tensorflow>=2.15
streamlit>=1.28
pillow>=10.0
jupyter>=1.0.0
